pyproject.toml
[project]
name = "gradingsystembackend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.8",
    "granian>=1.7.6",
    "pydantic-settings>=2.8.0",
    "sqlmodel>=0.0.22",
]

[dependency-groups]
dev = [
    "ruff>=0.9.7",
]

[tool.ruff]
line-length = 120
target-version = "py311"
lint.select = [
    "ALL",
]
lint.ignore = [
    "D",
    "COM812",
    "FBT001",
    "RUF001",
    "PLR0913",
    "TC002",
    "TRY003",
    "EM101",
    "DTZ005",
]
lint.fixable = [
    "F401", # delete unused imports
    "I001", # sort imports
    "W292", # newline at end of file
    "RUF022", # sort __all__ list
]
